🔄 [MOD] (core) Improve relation API (replace with a class).
